Bugfix for CameraMain for saving photos
Bug where if CameraMain.captureStillPicture was called too quickly the txt files would be saved with the wrong timestamp. Now added boolean flag (CameraMain.takingPhoto) which only allow photos to be taken once the current photo has finished being taken. captureStillPicture now returns boolean to indicate whether photo was taken successfully or not. This means captureStillPicture will not always take a photo, so tasks need to be able to handle this behaviour.